pyproject.toml
[project]
name = "sentinel-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "fastmcp>=3.2.3",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-mistralai>=1.1.2",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
